separated cli from api code to allow other projects to leverage the api directly if needed.
